Modularize role
Splitting the role in smaller parts helps to keep the overview and to
develop separate tests.
