+++ module_utils/network_lsr/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/python3 -tt
+# vim: fileencoding=utf8
+# SPDX-License-Identifier: BSD-3-Clause
+
+
+class MyError(Exception):
+    pass
