+++ tox.ini
@@ -8,15 +8,16 @@ basepython = python2.7
 deps =
     py{26,27,36,37}: pytest-cov
     py{27,36,37}: pytest>=3.5.1
+    py{26,27}: mock
     py26: pytest
 
 [base]
 passenv = *
 setenv =
-    PYTHONPATH = {toxinidir}/library
+    PYTHONPATH = {toxinidir}/library:{toxinidir}/module_utils
     LC_ALL = C
 changedir = {toxinidir}/tests
-covtarget = network_connections
+covtarget = {toxinidir}/library --cov {toxinidir}/module_utils
 pytesttarget = .
 
 [testenv:black]
@@ -93,6 +94,7 @@ commands =
         --errors-only \
         {posargs} \
         library/network_connections.py \
+        module_utils/network_lsr \
         tests/test_network_connections.py
 
 [testenv:flake8]
